requirements.txt
keyboard==0.13.5
matplotlib==3.8.3
mediapipe==0.10.11
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
pandas==2.2.1
pillow==10.2.0
pynput==1.7.6
pyserial==3.5
